renaming */example/ to */goodlisteners/ | pretty sure i'm breaking loads of shit.
