+++ pom.xml
@@ -2,7 +2,7 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.example</groupId>
+    <groupId>com.goodlisteners</groupId>
     <artifactId>my-java-project</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
